+++ HISTORY.md
@@ -1,13 +1,18 @@
 # History
 
 ## 23.2.0 (UNRELEASED)
+
 - Use [PDM](https://pdm.fming.dev/latest/) instead of Poetry.
 - _cattrs_ is now linted with [Ruff](https://beta.ruff.rs/docs/).
+- Fix TypedDicts with periods in their field names.
+  ([#376](https://github.com/python-attrs/cattrs/issues/376) [#377](https://github.com/python-attrs/cattrs/pull/377))
 
 ## 23.1.2 (2023-06-02)
+
 - Improve `typing_extensions` version bound. ([#372](https://github.com/python-attrs/cattrs/issues/372))
 
 ## 23.1.1 (2023-05-30)
+
 - Add `typing_extensions` as a direct dependency on 3.10. ([#369](https://github.com/python-attrs/cattrs/issues/369) [#370](https://github.com/python-attrs/cattrs/pull/370))
 
 ## 23.1.0 (2023-05-30)